doc: README, data1, fims_modules
* Add documentation about dataset source and format in data-raw folder
* Adds example for initialize_modules
* Adds minimal demo to README and removes vignette
* Adds cheatsheet to README
* Removes lots of partial documentation from fims_modules in favor of
  providing a link to the doxygen website for each function or class
* Ensure that all test @description are only one line and end in full stop
* Add lots of variables to utils::globalVariables
* Ignore some files from Rbuild
* Adds all Rcpp classes and functions to roxygen
* Uses links in doxygen and two groups to document the classes and functions

Part of #1188
Close #1174

+++ inst/include/distributions/functors/multinomial_lpmf.hpp
@@ -97,7 +97,9 @@ struct MultinomialLPMF : public DensityComponentBase<Type> {
         if (this->input_type == "data") {
           // if data, check if there are any NA values and skip lpdf calculation
           // for entire row if there are
-          if (this->get_observed(i, j) == this->observed_values->na_value) {
+          if (this->get_observed(static_cast<size_t>(i),
+                                 static_cast<size_t>(j)) ==
+              this->observed_values->na_value) {
             containsNA = true;
             break;
           }
